requirements.txt
psycopg2-binary==2.9.5
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0